add mirror

+++ crates/core/src/package/formula.rs
@@ -3,128 +3,170 @@ use serde::{Serialize, Deserialize};
 use serde_with::{serde_as, TryFromInto};
 
 // {
-//   "name": "postgresql@14",
-//   "full_name": "postgresql@14",
+//   "name": "postgresql@16",
+//   "full_name": "postgresql@16",
 //   "tap": "homebrew/core",
-//   "oldname": "postgresql",
+//   "oldname": null,
+//   "oldnames": [],
 //   "aliases": [],
-//   "versioned_formulae": ["postgresql@13", "postgresql@12", "postgresql@11", "postgresql@10", "postgresql@9.6", "postgresql@9.5", "postgresql@9.4"],
+//   "versioned_formulae": [
+//     "postgresql@15",
+//     "postgresql@14",
+//     "postgresql@13",
+//     "postgresql@12",
+//     "postgresql@11",
+//     "postgresql@10"
+//   ],
 //   "desc": "Object-relational database system",
 //   "license": "PostgreSQL",
 //   "homepage": "https://www.postgresql.org/",
 //   "versions": {
-//     "stable": "14.5",
+//     "stable": "16.2",
 //     "head": null,
 //     "bottle": true
 //   },
 //   "urls": {
 //     "stable": {
-//       "url": "https://ftp.postgresql.org/pub/source/v14.5/postgresql-14.5.tar.bz2",
+//       "url": "https://ftp.postgresql.org/pub/source/v16.2/postgresql-16.2.tar.bz2",
 //       "tag": null,
-//       "revision": null
+//       "revision": null,
+//       "using": null,
+//       "checksum": "446e88294dbc2c9085ab4b7061a646fa604b4bec03521d5ea671c2e5ad9b2952"
 //     }
 //   },
-//   "revision": 5,
+//   "revision": 1,
 //   "version_scheme": 0,
 //   "bottle": {
 //     "stable": {
 //       "rebuild": 0,
 //       "root_url": "https://ghcr.io/v2/homebrew/core",
 //       "files": {
+//         "arm64_sonoma": {
+//           "cellar": "/opt/homebrew/Cellar",
+//           "url": "https://ghcr.io/v2/homebrew/core/postgresql/16/blobs/sha256:40d2efe6bbcf70078f6fbe80ca8e72eafc16d6003d3903d475856f53514624a1",
+//           "sha256": "40d2efe6bbcf70078f6fbe80ca8e72eafc16d6003d3903d475856f53514624a1"
+//         },
+//         "arm64_ventura": {
+//           "cellar": "/opt/homebrew/Cellar",
+//           "url": "https://ghcr.io/v2/homebrew/core/postgresql/16/blobs/sha256:656a32a5ed4e7f8505f39d1e9e3b2b25ae2f153d7f25d774a87eb2b5fe8371d0",
+//           "sha256": "656a32a5ed4e7f8505f39d1e9e3b2b25ae2f153d7f25d774a87eb2b5fe8371d0"
+//         },
 //         "arm64_monterey": {
 //           "cellar": "/opt/homebrew/Cellar",
-//           "url": "https://ghcr.io/v2/homebrew/core/postgresql/14/blobs/sha256:5cb8cbf8fee5ba9a32a223deff80f6c885deb430d1bd38bc113e971d239d1534",
-//           "sha256": "5cb8cbf8fee5ba9a32a223deff80f6c885deb430d1bd38bc113e971d239d1534"
+//           "url": "https://ghcr.io/v2/homebrew/core/postgresql/16/blobs/sha256:c9d66495302339354a77a85caae0ab1cd579e4c218f3697c3dde2d438fe2edf4",
+//           "sha256": "c9d66495302339354a77a85caae0ab1cd579e4c218f3697c3dde2d438fe2edf4"
 //         },
-//         "arm64_big_sur": {
-//           "cellar": "/opt/homebrew/Cellar",
-//           "url": "https://ghcr.io/v2/homebrew/core/postgresql/14/blobs/sha256:52eabb4213030febe2435da260983eabbe7c6d29c2758829c53a43a8eb05d85c",
-//           "sha256": "52eabb4213030febe2435da260983eabbe7c6d29c2758829c53a43a8eb05d85c"
+//         "sonoma": {
+//           "cellar": "/usr/local/Cellar",
+//           "url": "https://ghcr.io/v2/homebrew/core/postgresql/16/blobs/sha256:82f8aa9bb1711ab710664f8b90d681c954a4b5b255203fab54db51a65fcd3715",
+//           "sha256": "82f8aa9bb1711ab710664f8b90d681c954a4b5b255203fab54db51a65fcd3715"
+//         },
+//         "ventura": {
+//           "cellar": "/usr/local/Cellar",
+//           "url": "https://ghcr.io/v2/homebrew/core/postgresql/16/blobs/sha256:d9a631c87687f289b61e6066808eb97a38dde76e61d97b51ed9f99fdae9d4538",
+//           "sha256": "d9a631c87687f289b61e6066808eb97a38dde76e61d97b51ed9f99fdae9d4538"
 //         },
 //         "monterey": {
 //           "cellar": "/usr/local/Cellar",
-//           "url": "https://ghcr.io/v2/homebrew/core/postgresql/14/blobs/sha256:f45e2b403e4dc0303f0c61cb429a1bd58c3c1cbe365adf5dbfada569cd2ba094",
-//           "sha256": "f45e2b403e4dc0303f0c61cb429a1bd58c3c1cbe365adf5dbfada569cd2ba094"
-//         },
-//         "big_sur": {
-//           "cellar": "/usr/local/Cellar",
-//           "url": "https://ghcr.io/v2/homebrew/core/postgresql/14/blobs/sha256:49a0e979e4c58f1e6e6c63b3ac118a4c9216428ae1bd3b1114ed24602b4dadda",
-//           "sha256": "49a0e979e4c58f1e6e6c63b3ac118a4c9216428ae1bd3b1114ed24602b4dadda"
-//         },
-//         "catalina": {
-//           "cellar": "/usr/local/Cellar",
-//           "url": "https://ghcr.io/v2/homebrew/core/postgresql/14/blobs/sha256:1d800b98a3a1715ec852beeb8fc6b28520d1f54f91ee521714309cfbfe139e2b",
-//           "sha256": "1d800b98a3a1715ec852beeb8fc6b28520d1f54f91ee521714309cfbfe139e2b"
+//           "url": "https://ghcr.io/v2/homebrew/core/postgresql/16/blobs/sha256:d57943164297ec488d94b5cacdbca72dc3f82048734185a127b7f609245f231a",
+//           "sha256": "d57943164297ec488d94b5cacdbca72dc3f82048734185a127b7f609245f231a"
 //         },
 //         "x86_64_linux": {
 //           "cellar": "/home/linuxbrew/.linuxbrew/Cellar",
-//           "url": "https://ghcr.io/v2/homebrew/core/postgresql/14/blobs/sha256:a4bf3d8e17f8a922d560c9d4f2151bc61f799d9dd15e29b1e9fc72fa9c6c4b12",
-//           "sha256": "a4bf3d8e17f8a922d560c9d4f2151bc61f799d9dd15e29b1e9fc72fa9c6c4b12"
+//           "url": "https://ghcr.io/v2/homebrew/core/postgresql/16/blobs/sha256:4655a82d8c2e9503f55bb453d71b28e313f81e78908670b19b8e741060ed02f4",
+//           "sha256": "4655a82d8c2e9503f55bb453d71b28e313f81e78908670b19b8e741060ed02f4"
 //         }
 //       }
 //     }
 //   },
-//   "keg_only": false,
-//   "keg_only_reason": null,
+//   "pour_bottle_only_if": null,
+//   "keg_only": true,
+//   "keg_only_reason": {
+//     "reason": ":versioned_formula",
+//     "explanation": ""
+//   },
 //   "options": [],
-//   "build_dependencies": ["pkg-config"],
-//   "dependencies": ["icu4c", "krb5", "lz4", "openssl@1.1", "readline"],
+//   "build_dependencies": [
+//     "pkg-config"
+//   ],
+//   "dependencies": [
+//     "gettext",
+//     "icu4c",
+//     "krb5",
+//     "lz4",
+//     "openssl@3",
+//     "readline",
+//     "zstd"
+//   ],
 //   "test_dependencies": [],
 //   "recommended_dependencies": [],
 //   "optional_dependencies": [],
-//   "uses_from_macos": ["libxml2", "libxslt", "openldap", "perl"],
+//   "uses_from_macos": [
+//     "libxml2",
+//     "libxslt",
+//     "openldap",
+//     "perl"
+//   ],
+//   "uses_from_macos_bounds": [
+//     {},
+//     {},
+//     {},
+//     {}
+//   ],
 //   "requirements": [],
 //   "conflicts_with": [],
-//   "caveats": "This formula has created a default database cluster with:\n  initdb --locale=C -E UTF-8 $(brew --prefix)/var/postgresql@14\nFor more details, read:\n  https://www.postgresql.org/docs/14/app-initdb.html\n",
-//   "installed": [{
-//     "version": "14.5_4",
-//     "used_options": [],
-//     "built_as_bottle": true,
-//     "poured_from_bottle": true,
-//     "time": 1664768139,
-//     "runtime_dependencies": [{
-//       "full_name": "icu4c",
-//       "version": "71.1",
-//       "declared_directly": true
-//     }, {
-//       "full_name": "ca-certificates",
-//       "version": "2022-07-19",
-//       "declared_directly": false
-//     }, {
-//       "full_name": "openssl@1.1",
-//       "version": "1.1.1q",
-//       "declared_directly": true
-//     }, {
-//       "full_name": "krb5",
-//       "version": "1.20",
-//       "declared_directly": true
-//     }, {
-//       "full_name": "lz4",
-//       "version": "1.9.4",
-//       "declared_directly": true
-//     }, {
-//       "full_name": "readline",
-//       "version": "8.1.2",
-//       "declared_directly": true
-//     }],
-//     "installed_as_dependency": false,
-//     "installed_on_request": true
-//   }],
-//   "linked_keg": "14.5_4",
+//   "conflicts_with_reasons": [],
+//   "link_overwrite": [],
+//   "caveats": "This formula has created a default database cluster with:\n  initdb --locale=C -E UTF-8 $HOMEBREW_PREFIX/var/postgresql@16\nFor more details, read:\n  https://www.postgresql.org/docs/16/app-initdb.html\n",
+//   "installed": [],
+//   "linked_keg": null,
 //   "pinned": false,
-//   "outdated": true,
+//   "outdated": false,
 //   "deprecated": false,
-//   "deprecation_date": "2026-11-12",
+//   "deprecation_date": "2028-11-09",
 //   "deprecation_reason": null,
 //   "disabled": false,
 //   "disable_date": null,
 //   "disable_reason": null,
+//   "post_install_defined": true,
+//   "service": {
+//     "run": [
+//       "$HOMEBREW_PREFIX/opt/postgresql@16/bin/postgres",
+//       "-D",
+//       "$HOMEBREW_PREFIX/var/postgresql@16"
+//     ],
+//     "run_type": "immediate",
+//     "keep_alive": {
+//       "always": true
+//     },
+//     "environment_variables": {
+//       "LC_ALL": "C"
+//     },
+//     "working_dir": "$HOMEBREW_PREFIX",
+//     "log_path": "$HOMEBREW_PREFIX/var/log/postgresql@16.log",
+//     "error_log_path": "$HOMEBREW_PREFIX/var/log/postgresql@16.log"
+//   },
+//   "tap_git_head": "0bbac30cc6051545a4b0d96ddd0289e6a23c4141",
+//   "ruby_source_path": "Formula/p/postgresql@16.rb",
+//   "ruby_source_checksum": {
+//     "sha256": "af43f12f1a9dc74b209c874db27a715bcbbeac4f6f30aa157b7a8b98e0926474"
+//   },
 //   "variations": {
 //     "x86_64_linux": {
-//       "dependencies": ["icu4c", "krb5", "lz4", "openssl@1.1", "readline", "libxml2", "libxslt", "openldap", "perl", "linux-pam", "util-linux"]
+//       "dependencies": [
+//         "gettext",
+//         "icu4c",
+//         "krb5",
+//         "lz4",
+//         "openssl@3",
+//         "readline",
+//         "zstd",
+//         "linux-pam",
+//         "util-linux"
+//       ]
 //     }
 //   }
-// },
+// }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Versions {
@@ -158,7 +200,7 @@ pub struct Bottle {
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Bottles {
-  pub rebuild: usize,
+  pub rebuild: u32,
   pub root_url: String,
   /// possible keys:
   /// arm64_sonoma, arm64_ventura, arm64_monterey,
@@ -176,8 +218,6 @@ pub type Dependencies = Vec<String>;
 pub enum RequirementName {
   Arch { version: String },
   Linux, Macos,
-  #[serde(alias = "maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_macos")]
-  #[serde(alias = "maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_maximum_macos")]
   MaximumMacos { version: String },
   Xcode,
   // this two only for glibc on linux
